[Update] close.js see if there is a role

+++ commands/close.js
@@ -10,22 +10,24 @@ module.exports = {
     async execute(message, args) {
         const role = message.member.roles.cache.find(role => role.name === message.channel.name)
 
-        // check channel
-        if (channels.includes(message.channel.name)) {
-            message.channel.send('You can\'t delete the this channel')
+        if (role) {
+            // check channel
+            if (channels.includes(message.channel.name)) {
+                message.channel.send('You can\'t delete the this channel')
 
-            // check member has role
-        } else if (message.member.roles.cache.has(role.id)) {
-            const millisecondsSleep = 10000
+                // check member has role
+            } else if (message.member.roles.cache.has(role.id)) {
+                const millisecondsSleep = 10000
 
-            message.channel.send('That\'s enough for today lets watch some interdimensional cable');
-            await sleep(millisecondsSleep)
+                message.channel.send('That\'s enough for today lets watch some interdimensional cable');
+                await sleep(millisecondsSleep)
 
-            role.delete()
-            message.channel.delete().catch(console.log)
+                role.delete()
+                message.channel.delete().catch(console.log)
 
-        } else {
-            message.channel.send('You don\'t have the right role to delete this channel')
+            } else {
+                message.channel.send('You don\'t have the right role to delete this channel')
+            }
         }
     }
 }
